Add files via upload
Thanks for your time last night and the review of the ChangePoint data to be checked and updated to drive accurate migration to Kimble.  An open question with regard to the new role assignment (Delivery Type) is given below.  Please also therefore update the Cost Role in line with these parameters.
The guidance for assigning was based around HR levels for headcount and used contractor feedback for resource managers, otherwise it was matched against the daily rates to
Junior – P1 – P2
Technical Consultant (etc.) – P3 – P4
Senior P5 - P6
For Kimble, as Rebekah mentioned on the call, the resource types will in future be :
Employee
Contractor
Consultant.  T&M Consultants will be a rarity within the CWF program in future, so we would expect to see this role as an exception in Kimble going forward.  They would trend as Co-Providers (outcome based, fixed fee), reporting to the Partner Management group, although their day-to-day deliverables would be driven by the requirements of the PMs / RMs.
Please take the time and opportunity to check other data in the various fields, such as resource type (especially for recent conversions to Employee, which might not have been updated in CP), Oracle Employee ID, cost centre and Reports To manager name.
Your kind attention to this update is appreciated, as well as your patience with the multiple times you have been requested to complete the exercise – ChangePoint updates have not always been effective, especially exits/off-boarding, as discussed on the call.
If you require any further information, please don’t hesitate to contact me.
Kind regards
Jo